backend/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
pluggy==1.3.0
puremagic==1.15
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyzbar==0.1.9
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
